requirements.txt
streamlit
openai
langchain
pillow
imageio
requests
